FEAT: SEO 디테일 페이지 추가
